+++ CHANGELOG.md
@@ -9,13 +9,16 @@ The format is based on the [KeepAChangeLog] project.
 ### Fixed
 - [#313]: Catch exception correctly
 - [#319]: Fix sanitize on strings starting with "B" or "U"
+- [#330]: Fix client_management user input being eval'd under Python 2
 
 ### Changed
 - [#318]: `oic.utils.authn.saml` raises `ImportError` on import if optional `saml2` dependency is not present.
+- [#325]: `oic.oic.claims_match` implementation refactored.
 
 [#313]: https://github.com/OpenIDC/pyoidc/issues/313
 [#318]: https://github.com/OpenIDC/pyoidc/pull/318
 [#319]: https://github.com/OpenIDC/pyoidc/pull/319
+[#325]: https://github.com/OpenIDC/pyoidc/pull/325
 
 ## 0.10.0.0 [2017-03-28]